// src/app/Models/OpDaily.ts
export class getDAD
{
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDAG {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNTSO {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNTIWB {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNTDie {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNTGas {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNTOWB {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDAWSO {
  Item1: number | undefined;
  Item2: Date | undefined;
  
}
export class getDAWIWB {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDAWDie {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDAWGas {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDAWOWB {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNBSO {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNBIWB {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNBDie {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNBGas {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNBOWB {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNFSO {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNFIWB {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNFDie {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNFGas {
  Item1: number | undefined;
  Item2: Date | undefined;

}
export class getDNFOWB {
  Item1: number | undefined;
  Item2: Date | undefined;

}